Get broadcasts

Get access to structured trasncripts data from broadcasts.

wh_broadcasts(token, q, ts = (Sys.time() - (3 * 24 * 60 * 60)),
  highlight = NULL, latest = NULL, quiet = !interactive())

Arguments

token
your token as returned by wh_token.

q
a string query containing the filters that define which transcript lines will be returned.

ts
The "ts" (timestamp) parameter is telling the system to return results that were crawled after this timestamp (POSIXct or POSIXlt).

highlight
return the fragments in the review that matched the textual Boolean query. The matched keywords will be surrounded by <em/> tags.

latest
this will return the latest 100 crawled posts matching your query (**NOT** recommended).

quiet
if FALSE does not return useful information to console.

Details

See official documentation for valid filters.

Examples

# NOT RUN {
token <- wh_token("xXX-x0X0xX0X-00X")

wef <- wh_broadcasts(token, q = '"World Economic Forum"') %>%  # use quote marks!
  wh_collect()

token %>%
  wh_broadcasts(token, q = 'Davos OR "World Economic Forum" OR WEF') %>%
  wh_paginate(p = 2) # 2 additional pages of results
  wh_collect() -> davos
# }